Move logic handling follows to separate file
